requirements.txt
requests==2.31.0
python-binance==1.0.19
python-telegram-bot==20.7
